+++ packages/kit/src/runtime/client/start.js
@@ -27,10 +27,6 @@ import { set_paths } from '../paths.js';
  * }} opts
  */
 export async function start({ paths, target, session, route, spa, trailing_slash, hydrate }) {
-	if (import.meta.env.DEV && !target) {
-		throw new Error('Missing target element. See https://kit.svelte.dev/docs#configuration-target');
-	}
-
 	const renderer = new Renderer({
 		Root,
 		fallback,