+++ packages/kit/src/utils/escape.js
@@ -1,50 +1,30 @@
+// dict from https://github.com/yahoo/serialize-javascript/blob/183c18a776e4635a379fdc620f81771f219832bb/index.js#L25
 /** @type {Record<string, string>} */
-const escape_json_string_in_html_dict = {
-	'"': '\\"',
+const escape_json_in_html_dict = {
 	'<': '\\u003C',
 	'>': '\\u003E',
 	'/': '\\u002F',
-	'\\': '\\\\',
-	'\b': '\\b',
-	'\f': '\\f',
-	'\n': '\\n',
-	'\r': '\\r',
-	'\t': '\\t',
-	'\0': '\\0',
 	'\u2028': '\\u2028',
 	'\u2029': '\\u2029'
 };
 
-/** @param {string} str */
-export function escape_json_string_in_html(str) {
-	return escape(
-		str,
-		escape_json_string_in_html_dict,
-		(code) => `\\u${code.toString(16).toUpperCase()}`
+const escape_json_in_html_regex = new RegExp(
+	`[${Object.keys(escape_json_in_html_dict).join('')}]`,
+	'g'
+);
+
+/**
+ * Escape a JSONValue that's going to be embedded in a `<script>` tag
+ * @param {import("@sveltejs/kit/types/helper").JSONValue} val
+ */
+export function escape_json_in_html(val) {
+	return JSON.stringify(val).replace(
+		escape_json_in_html_regex,
+		(match) => escape_json_in_html_dict[match]
 	);
 }
 
-/** @type {Record<string, string>} */
-const escape_html_attr_dict = {
-	'<': '&lt;',
-	'>': '&gt;',
-	'"': '&quot;'
-};
-
 /**
- * use for escaping string values to be used html attributes on the page
- * e.g.
- * <script data-url="here">
- *
- * @param {string} str
- * @returns string escaped string
- */
-export function escape_html_attr(str) {
-	return '"' + escape(str, escape_html_attr_dict, (code) => `&#${code};`) + '"';
-}
-
-/**
- *
  * @param str {string} string to escape
  * @param dict {Record<string, string>} dictionary of character replacements
  * @param unicode_encoder {function(number): string} encoder to use for high unicode characters
@@ -75,4 +55,23 @@ function escape(str, dict, unicode_encoder) {
 	}
 
 	return result;
+}
+
+/** @type {Record<string, string>} */
+const escape_html_attr_dict = {
+	'<': '&lt;',
+	'>': '&gt;',
+	'"': '&quot;'
+};
+
+/**
+ * use for escaping string values to be used html attributes on the page
+ * e.g.
+ * <script data-url="here">
+ *
+ * @param {string} str
+ * @returns string escaped string
+ */
+export function escape_html_attr(str) {
+	return '"' + escape(str, escape_html_attr_dict, (code) => `&#${code};`) + '"';
 }