+++ packages/kit/test/apps/basics/test/test.js
@@ -53,7 +53,7 @@ test.describe.parallel('a11y', () => {
 				page.type('#input', 'bar'),
 				page.waitForFunction(() => window.location.search === '?foo=bar')
 			]);
-			expect(await page.$eval('#input', (el) => el === document.activeElement)).toBe(true);
+			expect(await page.locator('#input')).toBeFocused();
 		}
 	});
 
@@ -178,7 +178,7 @@ test.describe('Scrolling', () => {
 	}) => {
 		await page.goto('/anchor');
 		await clicknav('#third-anchor');
-		expect(await page.evaluate(() => pageYOffset === 0)).toBeTruthy();
+		expect(await page.evaluate(() => scrollY === 0)).toBeTruthy();
 	});
 
 	test('url-supplied anchor works when navigated from bottom of page', async ({
@@ -197,7 +197,28 @@ test.describe('Scrolling', () => {
 	}) => {
 		await page.goto('/anchor');
 		await clicknav('#last-anchor-2');
-		expect(await page.evaluate(() => pageYOffset === 0)).toBeTruthy;
+		expect(await page.evaluate(() => scrollY === 0)).toBeTruthy();
+	});
+
+	test('scroll is restored after hitting the back button', async ({
+		back,
+		baseURL,
+		clicknav,
+		page
+	}) => {
+		await page.goto('/anchor');
+		await page.click('#scroll-anchor');
+		const originalScrollY = /** @type {number} */ (await page.evaluate(() => scrollY));
+		await clicknav('#routing-page');
+		await back();
+		expect(page.url()).toBe(baseURL + '/anchor#last-anchor-2');
+		const scrollY = /** @type {number} */ (await page.evaluate(() => scrollY));
+		expect(scrollY).toEqual(originalScrollY);
+		// TODO: fix this. it is failing due to duplicate history entries
+		// https://github.com/sveltejs/kit/issues/3636
+		// await page.goBack();
+		// expect(page.url()).toBe(baseURL + '/anchor');
+		// expect(scrollY).toEqual(0);
 	});
 
 	test('url-supplied anchor is ignored with onMount() scrolling on direct page load', async ({
@@ -221,7 +242,7 @@ test.describe('Scrolling', () => {
 	test('app-supplied scroll and focus work on direct page load', async ({ page, in_view }) => {
 		await page.goto('/use-action/focus-and-scroll');
 		expect(await in_view('#input')).toBe(true);
-		expect(await page.$eval('#input', (el) => el === document.activeElement)).toBe(true);
+		expect(await page.locator('#input')).toBeFocused();
 	});
 
 	test('app-supplied scroll and focus work on navigation to page', async ({
@@ -232,7 +253,7 @@ test.describe('Scrolling', () => {
 		await page.goto('/use-action');
 		await clicknav('[href="/use-action/focus-and-scroll"]');
 		expect(await in_view('#input')).toBe(true);
-		expect(await page.$eval('#input', (el) => el === document.activeElement)).toBe(true);
+		expect(await page.locator('#input')).toBeFocused();
 	});
 });
 
@@ -359,10 +380,32 @@ test.describe.parallel('Shadowed pages', () => {
 		expect(await page.textContent('h1')).toBe('Redirection was successful');
 	});
 
+	test('Handles GET redirects with cookies', async ({ page, context, clicknav }) => {
+		await page.goto('/shadowed');
+		await clicknav('[href="/shadowed/redirect-get-with-cookie"]');
+		expect(await page.textContent('h1')).toBe('Redirection was successful');
+
+		const cookies = await context.cookies();
+		expect(cookies).toEqual(
+			expect.arrayContaining([expect.objectContaining({ name: 'shadow-redirect', value: 'happy' })])
+		);
+	});
+
 	test('Handles POST redirects', async ({ page }) => {
 		await page.goto('/shadowed');
 		await Promise.all([page.waitForNavigation(), page.click('#redirect-post')]);
 		expect(await page.textContent('h1')).toBe('Redirection was successful');
+	});
+
+	test('Handles POST redirects with cookies', async ({ page, context }) => {
+		await page.goto('/shadowed');
+		await Promise.all([page.waitForNavigation(), page.click('#redirect-post-with-cookie')]);
+		expect(await page.textContent('h1')).toBe('Redirection was successful');
+
+		const cookies = await context.cookies();
+		expect(cookies).toEqual(
+			expect.arrayContaining([expect.objectContaining({ name: 'shadow-redirect', value: 'happy' })])
+		);
 	});
 
 	test('Renders error page for 4xx and 5xx responses from GET', async ({ page, clicknav }) => {
@@ -387,6 +430,44 @@ test.describe.parallel('Shadowed pages', () => {
 		});
 
 		expect(await response.json()).toEqual({ answer: 42 });
+	});
+
+	test('responds to HEAD requests from endpoint', async ({ request }) => {
+		const url = '/shadowed/simple';
+
+		const opts = {
+			headers: {
+				accept: 'application/json'
+			}
+		};
+
+		const responses = {
+			head: await request.head(url, opts),
+			get: await request.get(url, opts)
+		};
+
+		const headers = {
+			head: responses.head.headers(),
+			get: responses.get.headers()
+		};
+
+		expect(responses.head.status()).toBe(200);
+		expect(responses.get.status()).toBe(200);
+		expect(await responses.head.text()).toBe('');
+		expect(await responses.get.json()).toEqual({ answer: 42 });
+
+		['date', 'transfer-encoding'].forEach((name) => {
+			delete headers.head[name];
+			delete headers.get[name];
+		});
+
+		expect(headers.head).toEqual(headers.get);
+	});
+
+	test('Works with missing get handler', async ({ page, clicknav }) => {
+		await page.goto('/shadowed');
+		await clicknav('[href="/shadowed/no-get"]');
+		expect(await page.textContent('h1')).toBe('hello');
 	});
 });
 
@@ -414,6 +495,32 @@ test.describe.parallel('Endpoints', () => {
 		const response = await request.get('/endpoint-output/headers');
 		expect(/** @type {import('@playwright/test').APIResponse} */ (response).status()).toBe(200);
 		expect(response.headers()['set-cookie']).toBeDefined();
+	});
+
+	test('HEAD with matching headers but without body', async ({ request }) => {
+		const url = '/endpoint-output/body';
+
+		const responses = {
+			head: await request.head(url),
+			get: await request.get(url)
+		};
+
+		const headers = {
+			head: responses.head.headers(),
+			get: responses.get.headers()
+		};
+
+		expect(responses.head.status()).toBe(200);
+		expect(responses.get.status()).toBe(200);
+		expect(await responses.head.text()).toBe('');
+		expect(await responses.get.text()).toBe('{}');
+
+		['date', 'transfer-encoding'].forEach((name) => {
+			delete headers.head[name];
+			delete headers.get[name];
+		});
+
+		expect(headers.head).toEqual(headers.get);
 	});
 
 	test('200 status by default', async ({ request }) => {
@@ -740,6 +847,61 @@ test.describe.parallel('Errors', () => {
 		} else {
 			expect(contents).not.toMatch(location);
 		}
+	});
+
+	test('not ok response from endpoint', async ({ page, read_errors }) => {
+		const res = await page.goto('/errors/endpoint-not-ok');
+
+		expect(read_errors('/errors/endpoint-not-ok.json')).toBeUndefined();
+
+		expect(res && res.status()).toBe(555);
+		expect(await page.textContent('#message')).toBe('This is your custom error page saying: ""');
+
+		const contents = await page.textContent('#stack');
+		const location = 'endpoint-not-ok.svelte:12:15';
+
+		if (process.env.DEV) {
+			expect(contents).toMatch(location);
+		} else {
+			expect(contents).not.toMatch(location);
+		}
+	});
+
+	test('error in shadow endpoint', async ({ page, read_errors }) => {
+		const res = await page.goto('/errors/endpoint-shadow');
+
+		// should include stack trace
+		const lines = read_errors('/errors/endpoint-shadow').split('\n');
+		expect(lines[0]).toMatch('nope');
+
+		if (process.env.DEV) {
+			expect(lines[1]).toMatch('endpoint-shadow');
+		}
+
+		expect(res && res.status()).toBe(500);
+		expect(await page.textContent('#message')).toBe(
+			'This is your custom error page saying: "nope"'
+		);
+
+		const contents = await page.textContent('#stack');
+		const location = 'endpoint-shadow.js:1:8'; // TODO this is the wrong location, but i'm not going to open the sourcemap can of worms just now
+
+		if (process.env.DEV) {
+			expect(contents).toMatch(location);
+		} else {
+			expect(contents).not.toMatch(location);
+		}
+	});
+
+	test('not ok response from shadow endpoint', async ({ page, read_errors }) => {
+		const res = await page.goto('/errors/endpoint-shadow-not-ok');
+
+		expect(read_errors('/errors/endpoint-shadow-not-ok')).toBeUndefined();
+
+		expect(res && res.status()).toBe(555);
+		expect(await page.textContent('#message')).toBe(
+			'This is your custom error page saying: "Failed to load data"'
+		);
 	});
 
 	test('server-side 4xx status without error from load()', async ({ page }) => {
@@ -1133,7 +1295,7 @@ test.describe.parallel('Method overrides', () => {
 		await page.click('"No Override To GET"');
 
 		expect(await page.innerHTML('pre')).toBe(
-			'_method=GET is not allowed. See https://kit.svelte.dev/docs#configuration-methodoverride'
+			'_method=GET is not allowed. See https://kit.svelte.dev/docs/configuration#methodoverride'
 		);
 	});
 
@@ -1142,7 +1304,7 @@ test.describe.parallel('Method overrides', () => {
 		await page.click('"No Override To CONNECT"');
 
 		expect(await page.innerHTML('pre')).toBe(
-			'_method=CONNECT is not allowed. See https://kit.svelte.dev/docs#configuration-methodoverride'
+			'_method=CONNECT is not allowed. See https://kit.svelte.dev/docs/configuration#methodoverride'
 		);
 	});
 });
@@ -1262,6 +1424,11 @@ test.describe.parallel('Page options', () => {
 			await Promise.all([page.click('[href="/no-router/b"]'), page.waitForNavigation()]);
 			expect(await page.textContent('button')).toBe('clicks: 0');
 		}
+	});
+
+	test('transformPage can change the html output', async ({ page }) => {
+		await page.goto('/transform-page');
+		expect(await page.getAttribute('meta[name="transform-page"]', 'content')).toBe('Worked!');
 	});
 
 	test('does not SSR page with ssr=false', async ({ page, javaScriptEnabled }) => {
@@ -1483,14 +1650,16 @@ test.describe.parallel('Redirects', () => {
 	});
 
 	test('redirect-on-load', async ({ baseURL, page, javaScriptEnabled }) => {
-		await page.goto('/redirect-on-load');
+		const redirected_to_url = javaScriptEnabled
+			? `${baseURL}/redirect-on-load/redirected`
+			: `${baseURL}/redirect-on-load`;
+
+		await Promise.all([page.waitForResponse(redirected_to_url), page.goto('/redirect-on-load')]);
+
+		expect(page.url()).toBe(redirected_to_url);
 
 		if (javaScriptEnabled) {
-			await page.waitForTimeout(50); // TODO investigate why this test is flaky
-			expect(page.url()).toBe(`${baseURL}/redirect-on-load/redirected`);
 			expect(await page.textContent('h1')).toBe('Hazaa!');
-		} else {
-			expect(page.url()).toBe(`${baseURL}/redirect-on-load`);
 		}
 	});
 });
@@ -1627,9 +1796,21 @@ test.describe.parallel('Routing', () => {
 			let requests = [];
 			page.on('request', (r) => requests.push(r.url()));
 
-			await app.prefetch('/routing/prefetched');
-			expect(requests.length).toBe(2);
-			expect(requests[1]).toBe(`${baseURL}/routing/prefetched.json`);
+			// also wait for network processing to complete, see
+			// https://playwright.dev/docs/network#network-events
+			await Promise.all([
+				page.waitForResponse(`${baseURL}/routing/prefetched.json`),
+				app.prefetch('/routing/prefetched')
+			]);
+
+			// svelte request made is environment dependent
+			if (process.env.DEV) {
+				expect(requests.filter((req) => req.endsWith('index.svelte')).length).toBe(1);
+			} else {
+				expect(requests.filter((req) => req.endsWith('.js')).length).toBe(1);
+			}
+
+			expect(requests.includes(`${baseURL}/routing/prefetched.json`)).toBe(true);
 
 			requests = [];
 			await app.goto('/routing/prefetched');
@@ -1991,5 +2172,15 @@ test.describe.parallel('XSS', () => {
 
 		// @ts-expect-error - check global injected variable
 		expect(await page.evaluate(() => window.pwned)).toBeUndefined();
+	});
+
+	test('no xss via shadow endpoint', async ({ page }) => {
+		await page.goto('/xss/shadow');
+
+		// @ts-expect-error - check global injected variable
+		expect(await page.evaluate(() => window.pwned)).toBeUndefined();
+		expect(await page.textContent('h1')).toBe(
+			'user.name is </script><script>window.pwned = 1</script>'
+		);
 	});
 });