+++ packages/kit/types/index.d.ts
@@ -4,7 +4,21 @@
 import './ambient-modules';
 
 export { App, SSRManifest } from './app';
-export { Adapter, Builder, Config, PrerenderErrorHandler, ValidatedConfig } from './config';
+export {
+	Adapter,
+	Builder,
+	Config,
+	Prerendered,
+	PrerenderErrorHandler,
+	ValidatedConfig
+} from './config';
 export { EndpointOutput, RequestHandler } from './endpoint';
 export { ErrorLoad, ErrorLoadInput, Load, LoadInput, LoadOutput } from './page';
-export { ExternalFetch, GetSession, Handle, HandleError, RequestEvent, ResolveOpts } from './hooks';
+export {
+	ExternalFetch,
+	GetSession,
+	Handle,
+	HandleError,
+	RequestEvent,
+	ResolveOptions
+} from './hooks';