+++ sites/kit.svelte.dev/src/routes/docs/index.json.js
@@ -1,0 +1,7 @@
+import { read_all } from '$lib/docs';
+
+export function get() {
+	return {
+		body: read_all('docs').map(({ slug, title, sections }) => ({ slug, title, sections }))
+	};
+}