+++ sites/kit.svelte.dev/src/routes/faq/index.svelte
@@ -20,10 +20,9 @@
 	<h1>Frequently Asked Questions</h1>
 	{#each sections as faq}
 		<article class="faq">
-			<h2>
-				<span id={faq.slug} class="offset-anchor" />
+			<h2 id={faq.slug}>
 				{faq.title}
-				<Permalink href="/faq#{faq.slug}" />
+				<Permalink href="#{faq.slug}" />
 			</h2>
 			{@html faq.content}
 		</article>
@@ -38,41 +37,28 @@
 	.faqs {
 		grid-template-columns: 1fr 1fr;
 		grid-gap: 1em;
-		min-block-size: calc(100vh - var(--nav-h));
-		padding-inline: var(--side-nav);
-		padding-block: var(--top-offset) 6rem;
-		max-inline-size: var(--main-width);
-		margin-inline: auto;
-		margin-block: 0;
+		min-height: calc(100vh - var(--nav-h));
+		padding: var(--top-offset) var(--side-nav) 6rem var(--side-nav);
+		max-width: var(--main-width);
+		margin: 0 auto;
 		tab-size: 2;
 	}
 
 	.faqs :global(pre) :global(code) {
 		padding: 0;
 		margin: 0;
-		inset-block-start: 0;
+		top: 0;
 		background: transparent;
-		color: white;
 	}
 
 	.faqs :global(pre) {
 		margin: 0;
-		margin-block-end: 2rem;
-		inline-size: 100%;
-		max-inline-size: var(--linemax);
-		padding-inline: 2.5rem;
-		padding-block: 1.5rem;
-		background: #333;
+		margin-bottom: 2rem;
+		width: 100%;
+		max-width: var(--linemax);
+		padding: 1.5rem 2.5rem;
 		border-radius: 0.5rem;
 		font-size: 0.8rem;
-	}
-
-	.faqs :global(.offset-anchor) {
-		position: relative;
-		display: block;
-		inset-block-start: calc(-1 * var(--top-offset));
-		inline-size: 0;
-		block-size: 0;
 	}
 
 	.faqs :global(.anchor) {
@@ -80,16 +66,16 @@
 		display: block;
 		background: url(@sveltejs/site-kit/icons/link.svg) 0 50% no-repeat;
 		background-size: 1em 1em;
-		inline-size: 1.4em;
-		block-size: 1em;
-		inset-inline-start: -1.3em;
+		width: 1.4em;
+		height: 1em;
+		left: -1.3em;
 		opacity: 0;
 		transition: opacity 0.2s;
 	}
 
 	.faqs :global(h2 > .anchor),
 	.faqs :global(h3 > .anchor) {
-		inset-block-start: 0.2em;
+		bottom: 0.3em;
 	}
 
 	@media (min-width: 768px) {
@@ -104,18 +90,19 @@
 
 		.faqs :global(h5) :global(.anchor),
 		.faqs :global(h6) :global(.anchor) {
-			inset-block-start: 0.25em;
+			bottom: 0.25em;
 		}
 	}
 
 	h2 {
-		margin-block: 3.5rem 1rem;
-		padding-block-end: 0.2rem;
+		margin: -4rem 0 1rem 0;
+		padding-top: 10rem;
+		padding-bottom: 0.2rem;
 		color: var(--text);
-		/* max-inline-size: 24em; */
+		/* max-width: 24em; */
 		font-size: var(--h3);
 		font-weight: 400;
-		border-block-end: 1px solid #ddd;
+		border-bottom: 1px solid #ddd;
 	}
 
 	.faqs :global(h3) {
@@ -123,19 +110,28 @@
 		font-weight: 600;
 		font-size: 2rem;
 		color: var(--second);
-		margin-block: 2rem 1.6rem;
-		padding-inline-start: 0;
+		margin: 2rem 0 1.6rem 0;
+		padding-left: 0;
 		background: transparent;
 		line-height: 1.3;
 		padding: 0;
-		inset-block-start: 0;
+		top: 0;
+	}
+
+	.faqs :global(a) {
+		position: relative;
+		z-index: 2;
+	}
+
+	/* TODO this page must be missing some styles from somewhere. the whole thing needs tidying up */
+	.faqs :global(a) :global(code) {
+		color: var(--prime);
 	}
 
 	.faq:first-child {
-		margin: 0;
-		margin-block: 2rem;
-		padding-block-end: 4rem;
-		border-block-end: var(--border-w) solid #6767785b; /* based on --second */
+		margin: 2rem 0;
+		padding-bottom: 4rem;
+		border-bottom: var(--border-w) solid #6767785b; /* based on --second */
 	}
 	.faq:first-child h2 {
 		font-size: 4rem;
@@ -144,18 +140,14 @@
 	}
 
 	:global(.faqs .faq ul) {
-		margin-inline-start: 3.2rem;
-	}
-
-	.faqs :global(.anchor) {
-		inset-block-start: calc((var(--h3) - 24px) / 2);
+		margin-left: 3.2rem;
 	}
 
 	@media (max-width: 768px) {
 		.faqs :global(.anchor) {
 			transform: scale(0.6);
 			opacity: 1;
-			inset-inline-start: -1em;
+			left: -1em;
 		}
 	}
 </style>