+++ sites/kit.svelte.dev/src/routes/migrating/index.js
@@ -1,9 +1,0 @@
-import { read } from '$lib/docs';
-
-export function get() {
-	return {
-		body: {
-			sections: read('migrating')
-		}
-	};
-}