+++ packages/adapter-cloudflare/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sveltejs/adapter-cloudflare
 
+## 1.0.0-next.11
+
+### Patch Changes
+
+- Add `context` to `event.platform` object ([#3868](https://github.com/sveltejs/kit/pull/3868))
+
+## 1.0.0-next.10
+
+### Patch Changes
+
+- update to Vite 2.8 and esbuild 0.14 ([#3791](https://github.com/sveltejs/kit/pull/3791))
+
 ## 1.0.0-next.9
 
 ### Patch Changes